Don't unroll guards used by ellipses code gen
This on top of all of the other code gen is causing traces to get too
big, which sometimes causes the interpreter to fail silently. It
appears to be related to the JIT, since I can run some code that
doesn't get traced but once I cross the tracing threshold then the
silent failures come.