Use none term in main loop; Catch errors raised in main loop
The variable `tmp` in the main loop is so that we only assign a term
to c, a green variable, if it's static; without that we'd be
potentially promoting virtuals.

There are a couple hacky references to runtime bits---anywhere with a
module-var* or lexical-var* in to-py.rkt is likely one of those hacks.

In the main loop: accessing the exception's message caused a
translation error related to it being readonly, soo just try to format
the whole exception as a string