Use more approachable cont names; Factor out some apply-related bits
It would be nice to use the name "return" for the ret metafunction,
but at this point I don't sanitize those names before producing Python
code (the return keyword is an id according to Python's grammar, so
pythonify-name doesn't complain)

It might also be nice to allow the ret metafunction to come after the
transition metafunction. With the current setup that would require
lifting the define-transition to the end of the module, much like the
call to register-metafunction! is lifted