Run Racket programs on pycketlite with `--racket path`
I need to replace `jam-run` with two forms: one interactive and one
that takes a datum and returns whether or not executing it exited
under normal conditions. Then the tests added by this will depend on
the behavior of the abstract machine; currently they pass as long as
the call to jam-run happens (e.g. expansion failures mean test
failurse; error during jam-run though doesn't cause the test to fail)