Convert the Pycketlite transition function from eval to eval-apply
The new tests were part of me debugging a test failure that cropped
up. The test
    (test-equal (run-eval-e ((lambda (n) (+ 1 n)) 10)) 11)
was the simplest one that triggered the bug, and the fault was that
the unload pattern in the eval-e evaluator overlapped with the rules
that determine whether the current top-level form is an expression or
a definition form. Specifically, a length 2 application-expression
overlapped with the {_ v} pattern in the previous unload pattern.