Add values/call-with-values and multivalue continuations
Factoring continuations into two groups, single value and multiple
value, is analagous to how evaluation contexts in the R5RS operational
semantics are factored into two groups. The transitions that
wrap/unwrap #%values---(#%values V) returning to a k1 and V returning
to a k*---are my version of the VDemote and VPromote rule from that
semantics, too.