Add let-values
Racket's let-values checks that the number of values from a RHS
matches the number of variables on the corresponding LHS once the RHS
is fully.

There are a couple different ways to write a continuation for
let-values but having the LHS for the current RHS in a fixed spot in
the continuation, like this one does, makes it easier to check that
the number of values and variables match.