Remove some recursion from the pycketlite grammar, runtime
I could only get away with removing recursion from forms with tags,
like `if`. Since application forms aren't tagged, they overlap with
closures, and returning a closure to a topk gets misinterpreted as an
expression being evaluated at toplevel (which then evaluates the same
"expression" in the environment stored in the topk).

Checking list patterns is also a place where we recur on terms, but
that had to be fixed at the runtime-level