Omit Racket's begin-for-syntax in pycketlite
The pycket/wrap language inserts relative paths but doesn't set
current-load-relative-directory; I'm not sure why they don't but to
expand a benchmark from an arbitrary directory I had to set it. I set
it by expanding into a begin-for-syntax that sets the directory, and
so I need to omit begin-for-syntax.