WIP: Hack together callgraph support for Pycketlite
When I initialized the instance-surrounding_lambda in the constructor
I was getting that annoying assertion error in check_result about the
number of refs/constants. So now I have a class member and the
instance member otherwise gets initialized by calling
set_surrounding_lambda.

It's not great that I had to change callgraph.py. Maybe there's a
better way to do this when this goes from WIP -> implemented.

I also decided to make the appk return the next expression to evaluate
rather than the (nonempty) list of remaining expressions to
evaluate. I think this was necessary because for some reason the list
of remaining expressions didn't have the surrounding_lambda
set (probably because of reconstructing the list of expressions in the
app and appk rules)

Dynamic env thing